// skyvern-ts/client/src/environments.ts
// This file was auto-generated by Fern from our API Definition.

export const SkyvernEnvironment = {
    Cloud: "https://api.skyvern.com",
    Staging: "https://api-staging.skyvern.com",
    Local: "http://localhost:8000",
} as const;

export type SkyvernEnvironment =
    | typeof SkyvernEnvironment.Cloud
    | typeof SkyvernEnvironment.Staging
    | typeof SkyvernEnvironment.Local;